feat(supplier): add supplier icon
