+++ packages/ocean-icons-react/src/SupplierOutline.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+
+type IconProps = {
+  size?: number;
+} & React.SVGProps<SVGSVGElement>;
+
+const SupplierOutline = (
+  { size = 24, ...props }: IconProps,
+  svgRef: React.Ref<SVGSVGElement>
+) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+    viewBox="0 0 24 24"
+    stroke="currentColor"
+    width={size}
+    height={size}
+    ref={svgRef}
+    {...props}
+  >
+    <path d="M3.4 6.96667L4.2 7.56667L4.2 7.56667L3.4 6.96667ZM5.4 4.3L6.2 4.9L6.2 4.9L5.4 4.3ZM18.6 4.3L19.4 3.7L19.4 3.7L18.6 4.3ZM20.6 6.96667L21.4 6.36667L21.4 6.36667L20.6 6.96667ZM9 10.5C8.44772 10.5 8 10.9477 8 11.5C8 12.0523 8.44772 12.5 9 12.5V10.5ZM15 12.5C15.5523 12.5 16 12.0523 16 11.5C16 10.9477 15.5523 10.5 15 10.5V12.5ZM11 7.5C11 8.05228 11.4477 8.5 12 8.5C12.5523 8.5 13 8.05228 13 7.5H11ZM13 3.5C13 2.94772 12.5523 2.5 12 2.5C11.4477 2.5 11 2.94772 11 3.5H13ZM2 8.16403V19.5032H4V8.16403H2ZM5 22.5H19V20.5H5V22.5ZM22 19.5V8.16667H20V19.5H22ZM4.2 7.56667L6.2 4.9L4.6 3.7L2.6 6.36667L4.2 7.56667ZM7 4.5H17V2.5H7V4.5ZM17.8 4.9L19.8 7.56667L21.4 6.36667L19.4 3.7L17.8 4.9ZM17 4.5C17.3148 4.5 17.6111 4.64819 17.8 4.9L19.4 3.7C18.8334 2.94458 17.9443 2.5 17 2.5V4.5ZM6.2 4.9C6.38885 4.64819 6.68524 4.5 7 4.5V2.5C6.05573 2.5 5.16656 2.94458 4.6 3.7L6.2 4.9ZM22 8.16667C22 7.51756 21.7895 6.88596 21.4 6.36667L19.8 7.56667C19.9298 7.73976 20 7.9503 20 8.16667H22ZM19 22.5C20.6569 22.5 22 21.1569 22 19.5H20C20 20.0523 19.5523 20.5 19 20.5V22.5ZM2 19.5032C2 21.1621 3.3452 22.5 5 22.5V20.5C4.44566 20.5 4 20.0534 4 19.5032H2ZM4 8.16403C4 7.94903 4.06935 7.74087 4.2 7.56667L2.6 6.36667C2.21136 6.88485 2 7.51354 2 8.16403H4ZM4 8.5H20V6.5H4V8.5ZM9 12.5H15V10.5H9V12.5ZM13 7.5V3.5H11V7.5H13Z" />
+  </svg>
+);
+
+const ForwardRef = React.forwardRef(SupplierOutline);
+export default ForwardRef;
